requirements.txt
SQLAlchemy
nest-asyncio
requests
openai
uvicorn
numpy
pydantic
starlette
slowapi
typing-extensions
fastapi
passlib
python-jose
mysql-connector-python
pydantic[email]
bcrypt
xmltodict
APScheduler
firebase-admin
pytz
schedule
firebase_admin
dotenv
pandas
pymysql